Added Memory Message Bus
